Module zerovec::map

See ZeroMap for details.

Structs

ZeroMap

A zero-copy map datastructure, built on sorted binary-searchable ZeroVec and VarZeroVec.

ZeroMapBorrowed

A borrowed-only version of ZeroMap

Traits

BorrowedZeroVecLike

Trait abstracting over ZeroVec and VarZeroVec, for use in ZeroMap. You should not be implementing or calling this trait directly.

MutableZeroVecLike

Trait abstracting over ZeroVec and VarZeroVec, for use in ZeroMap. You should not be implementing or calling this trait directly.

ZeroMapKV

Trait marking types which are allowed to be keys or values in ZeroMap.

ZeroVecLike

Trait abstracting over ZeroVec and VarZeroVec, for use in ZeroMap. You should not be implementing or calling this trait directly.