Trait zerovec::map::ZeroVecLike

pub trait ZeroVecLike<'a, T: ?Sized> {
    type NeedleType: ?Sized;
    type GetType: ?Sized + 'static;
    fn new() -> Self;

    fn binary_search(&self, k: &Self::NeedleType) -> Result<usize, usize>;

    fn get(&self, index: usize) -> Option<&Self::GetType>;

    fn len(&self) -> usize;

    fn is_ascending(&self) -> bool;

    fn is_empty(&self) -> bool { ... }
}

Trait abstracting over ZeroVec and VarZeroVec, for use in ZeroMap. You should not be implementing or calling this trait directly.

Associated Types

type NeedleType: ?Sized

The type received by Self::binary_search()

type GetType: ?Sized + 'static

The type returned by Self::get()

Required methods

fn new() -> Self

Create a new, empty vector

fn binary_search(&self, k: &Self::NeedleType) -> Result<usize, usize>

Search for a key in a sorted vector, returns Ok(index) if found, returns Err(insert_index) if not found, where insert_index is the index where it should be inserted to maintain sort order.

fn get(&self, index: usize) -> Option<&Self::GetType>

Get element at index

fn len(&self) -> usize

The length of this vector

fn is_ascending(&self) -> bool

Check if this vector is in ascending order according to Ts Ord impl

Provided methods

fn is_empty(&self) -> bool

Check if this vector is empty