Struct zerovec::map2d::ZeroMap2d

pub struct ZeroMap2d<'a, K0, K1, V> where
    K0: ZeroMapKV<'a>,
    K1: ZeroMapKV<'a>,
    V: ZeroMapKV<'a>,
    K0: ?Sized,
    K1: ?Sized,
    V: ?Sized,  { /* fields omitted */ }

A zero-copy, two-dimensional map datastructure .

This is an extension of ZeroMap that supports two layers of keys. For example, to map a pair of an integer and a string to a buffer, you can write:

let _: ZeroMap2d<u32, str, [u8]> = unimplemented!();

Internally, ZeroMap2d stores four zero-copy vectors, one for each type argument plus one more to match between the two vectors of keys.

Examples

use zerovec::ZeroMap2d;

// Example byte buffer representing the map { 1: {2: "three" } }
let BINCODE_BYTES: &[u8; 53] = &[
    2, 0, 0, 0, 0, 0, 0, 0, 1, 0, 4, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 2, 0, 0, 0, 0, 0,
    0, 0, 2, 0, 13, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 116, 104, 114, 101, 101
];

// Deserializing to ZeroMap requires no heap allocations.
let zero_map: ZeroMap2d<u16, u16, str> = bincode::deserialize(BINCODE_BYTES)
    .expect("Should deserialize successfully");
assert_eq!(zero_map.get(&1, &2), Some("three"));

Implementations

impl<'a, K0, K1, V> ZeroMap2d<'a, K0, K1, V> where
K0: ZeroMapKV<'a>,
K1: ZeroMapKV<'a>,
V: ZeroMapKV<'a>,
K0: ?Sized,
K1: ?Sized,
V: ?Sized,

pub fn new() -> Self

Creates a new, empty ZeroMap2d.

Examples

use zerovec::ZeroMap2d;

let zm: ZeroMap2d<u16, str, str> = ZeroMap2d::new();
assert!(zm.is_empty());

pub fn with_capacity(capacity: usize) -> Self

Construct a new ZeroMap2d with a given capacity

pub fn as_borrowed(&'a self) -> ZeroMap2dBorrowed<'a, K0, K1, V>

Obtain a borrowed version of this map

pub fn len(&self) -> usize

The number of values in the ZeroMap2d

pub fn is_empty(&self) -> bool

Whether the ZeroMap2d is empty

pub fn clear(&mut self)

Remove all elements from the ZeroMap2d

pub fn reserve(&mut self, additional: usize)

Reserve capacity for additional more elements to be inserted into the ZeroMap2d to avoid frequent reallocations.

See Vec::reserve() for more information.

pub fn get(&self, key0: &K0, key1: &K1) -> Option<&V::GetType>

Get the value associated with key0 and key1, if it exists.

use zerovec::ZeroMap2d;

let mut map = ZeroMap2d::new();
map.insert(&1, "one", "foo");
map.insert(&2, "one", "bar");
map.insert(&2, "two", "baz");
assert_eq!(map.get(&1, "one"), Some("foo"));
assert_eq!(map.get(&1, "two"), None);
assert_eq!(map.get(&2, "one"), Some("bar"));
assert_eq!(map.get(&2, "two"), Some("baz"));
assert_eq!(map.get(&3, "three"), None);

pub fn contains_key0(&self, key0: &K0) -> bool

Returns whether key0 is contained in this map

use zerovec::ZeroMap2d;

let mut map = ZeroMap2d::new();
map.insert(&1, "one", "foo");
map.insert(&2, "two", "bar");
assert_eq!(map.contains_key0(&1), true);
assert_eq!(map.contains_key0(&3), false);

pub fn insert(
&mut self,
key0: &K0,
key1: &K1,
value: &V
) -> Option<V::OwnedType>

Insert value with key, returning the existing value if it exists.

See example in Self::get().

pub fn remove(&mut self, key0: &K0, key1: &K1) -> Option<V::OwnedType>

Remove the value at key, returning it if it exists.

use zerovec::ZeroMap2d;

let mut map = ZeroMap2d::new();
map.insert(&1, "one", "foo");
map.insert(&2, "two", "bar");
assert_eq!(map.remove(&1, "one"), Some("foo".to_owned().into_boxed_str()));
assert_eq!(map.get(&1, "one"), None);

pub fn try_append<'b>(
&mut self,
key0: &'b K0,
key1: &'b K1,
value: &'b V
) -> Option<(&'b K0, &'b K1, &'b V)>

Appends value with key to the end of the underlying vector, returning key and value if it failed. Useful for extending with an existing sorted list.

use zerovec::ZeroMap2d;

let mut map = ZeroMap2d::new();
assert!(map.try_append(&1, "one", "uno").is_none());
assert!(map.try_append(&3, "three", "tres").is_none());

let unsuccessful = map.try_append(&3, "three", "tres-updated");
assert!(unsuccessful.is_some(), "append duplicate of last key");

let unsuccessful = map.try_append(&2, "two", "dos");
assert!(unsuccessful.is_some(), "append out of order");

assert_eq!(map.get(&1, "one"), Some("uno"));

// contains the original value for the key: 3
assert_eq!(map.get(&3, "three"), Some("tres"));

// not appended since it wasn't in order
assert_eq!(map.get(&2, "two"), None);

pub fn iter_keys0<'b>(
&'b self
) -> impl Iterator<Item = &'b <K0 as ZeroMapKV<'a>>::GetType>

Produce an ordered iterator over keys0.

pub fn iter_keys1<'b>(
&'b self,
key0: &K0
) -> Option<impl Iterator<Item = &'b <K1 as ZeroMapKV<'a>>::GetType>>

Produce an ordered iterator over keys1 for a particular key0, if key0 exists.

pub fn iter_keys1_by_index<'b>(
&'b self,
key0_index: usize
) -> Option<impl Iterator<Item = &'b <K1 as ZeroMapKV<'a>>::GetType>>

Produce an ordered iterator over keys1 for a particular key0_index, if key0_index exists.

This method is designed to interoperate with the enumerated key of iter_keys0.

Panics

Panics if key0_index is out of range.

Example

Loop over all elements of a ZeroMap2d:

use zerovec::ZeroMap2d;

let mut map: ZeroMap2d<u16, u16, str> = ZeroMap2d::new();
map.insert(&1, &1, "foo");
map.insert(&2, &3, "bar");
map.insert(&2, &4, "baz");

let mut total_value = 0;

let mut values_it = map.iter_values();
for (key0_index, key0) in map.iter_keys0().enumerate() {
    for key1 in map.iter_keys1_by_index(key0_index).unwrap() {
        // This code runs for every (key0, key1) pair
        total_value += key0.as_unsigned_int() as usize;
        total_value += key1.as_unsigned_int() as usize;
        total_value += values_it.next().unwrap().len();
    }
}

assert_eq!(total_value, 22);

pub fn iter_values<'b>(
&'b self
) -> impl Iterator<Item = &'b <V as ZeroMapKV<'a>>::GetType>

Produce an iterator over values, ordered by the pair (key0,key1)

impl<'a, K0, K1, V> ZeroMap2d<'a, K0, K1, V> where
K0: ZeroMapKV<'a> + ?Sized,
K1: ZeroMapKV<'a> + ?Sized,
V: ZeroMapKV<'a, Container = ZeroVec<'a, V>> + ?Sized,
V: AsULE + Copy,

pub fn get_copied(&self, key0: &K0, key1: &K1) -> Option<V>

For cases when V is fixed-size, obtain a direct copy of V instead of V::ULE

Examples

let mut map: ZeroMap2d::<u16, u16, u16> = ZeroMap2d::new();
map.insert(&1, &2, &3);
map.insert(&1, &4, &5);
map.insert(&6, &7, &8);

assert_eq!(map.get_copied(&6, &7), Some(8));