Struct zerovec::maps::ZeroMapBorrowed

pub struct ZeroMapBorrowed<'a, K, V> where
    K: ZeroMapKV<'a>,
    V: ZeroMapKV<'a>,
    K: ?Sized,
    V: ?Sized,  { /* private fields */ }

A borrowed-only version of ZeroMap

This is useful for fully-zero-copy deserialization from non-human-readable serialization formats. It also has the advantage that it can return references that live for the lifetime of the backing buffer as opposed to that of the ZeroMapBorrowed instance.

Examples

use zerovec::maps::ZeroMapBorrowed;

// Example byte buffer representing the map { 1: "one" }
let BINCODE_BYTES: &[u8; 31] = &[
    4, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 11, 0, 0, 0, 0, 0, 0, 0,
    1, 0, 0, 0, 0, 0, 0, 0, 111, 110, 101
];

// Deserializing to ZeroMap requires no heap allocations.
let zero_map: ZeroMapBorrowed<u32, str> = bincode::deserialize(BINCODE_BYTES)
    .expect("Should deserialize successfully");
assert_eq!(zero_map.get(&1), Some("one"));

This can be obtained from a ZeroMap via ZeroMap::as_borrowed

Implementations

impl<'a, K, V> ZeroMapBorrowed<'a, K, V> where
K: ZeroMapKV<'a>,
V: ZeroMapKV<'a>,
K: ?Sized,
V: ?Sized,

pub fn new() -> Self

Creates a new, empty ZeroMapBorrowed<K, V>.

Note: Since ZeroMapBorrowed is not mutable, the return value will be a stub unless converted into a ZeroMap.

Examples

use zerovec::maps::ZeroMapBorrowed;

let zm: ZeroMapBorrowed<u16, str> = ZeroMapBorrowed::new();
assert!(zm.is_empty());

pub fn len(&self) -> usize

The number of elements in the ZeroMapBorrowed

pub fn is_empty(&self) -> bool

Whether the ZeroMapBorrowed is empty

impl<'a, K, V> ZeroMapBorrowed<'a, K, V> where
K: ZeroMapKV<'a> + Ord,
V: ZeroMapKV<'a>,
K: ?Sized,
V: ?Sized,

pub fn get(&self, key: &K) -> Option<&'a V::GetType>

Get the value associated with key, if it exists.

This is able to return values that live longer than the map itself since they borrow directly from the backing buffer. This is the primary advantage of using ZeroMapBorrowed over ZeroMap.

use zerovec::ZeroMap;
use zerovec::maps::ZeroMapBorrowed;

let mut map = ZeroMap::new();
map.insert(&1, "one");
map.insert(&2, "two");
let borrowed = map.as_borrowed();
assert_eq!(borrowed.get(&1), Some("one"));
assert_eq!(borrowed.get(&3), None);

let borrow = borrowed.get(&1);
drop(borrowed);
// still exists after the ZeroMapBorrowed has been dropped
assert_eq!(borrow, Some("one"));

pub fn get_by(
&self,
predicate: impl FnMut(&K) -> Ordering
) -> Option<&'a V::GetType>

Binary search the map with predicate to find a key, returning the value.

This is able to return values that live longer than the map itself since they borrow directly from the backing buffer. This is the primary advantage of using ZeroMapBorrowed over ZeroMap.

use zerovec::ZeroMap;
use zerovec::maps::ZeroMapBorrowed;

let mut map = ZeroMap::new();
map.insert(&1, "one");
map.insert(&2, "two");
let borrowed = map.as_borrowed();
assert_eq!(borrowed.get_by(|probe| probe.cmp(&1)), Some("one"));
assert_eq!(borrowed.get_by(|probe| probe.cmp(&3)), None);

let borrow = borrowed.get_by(|probe| probe.cmp(&1));
drop(borrowed);
// still exists after the ZeroMapBorrowed has been dropped
assert_eq!(borrow, Some("one"));

pub fn contains_key(&self, key: &K) -> bool

Returns whether key is contained in this map

use zerovec::ZeroMap;
use zerovec::maps::ZeroMapBorrowed;

let mut map = ZeroMap::new();
map.insert(&1, "one");
map.insert(&2, "two");
let borrowed = map.as_borrowed();
assert_eq!(borrowed.contains_key(&1), true);
assert_eq!(borrowed.contains_key(&3), false);

impl<'a, K, V> ZeroMapBorrowed<'a, K, V> where
K: ZeroMapKV<'a>,
V: ZeroMapKV<'a>,
K: ?Sized,
V: ?Sized,

pub fn iter<'b>(
&'b self
) -> impl Iterator<Item = (&'a <K as ZeroMapKV<'a>>::GetType, &'a <V as ZeroMapKV<'a>>::GetType)> + 'b

Produce an ordered iterator over key-value pairs

pub fn iter_keys<'b>(
&'b self
) -> impl Iterator<Item = &'a <K as ZeroMapKV<'a>>::GetType> + 'b

Produce an ordered iterator over keys

pub fn iter_values<'b>(
&'b self
) -> impl Iterator<Item = &'a <V as ZeroMapKV<'a>>::GetType> + 'b

Produce an iterator over values, ordered by keys

impl<'a, K, V> ZeroMapBorrowed<'a, K, V> where
K: ZeroMapKV<'a> + ?Sized + Ord,
V: ZeroMapKV<'a, Container = ZeroVec<'a, V>> + ?Sized,
V: AsULE + Ord + Copy + 'static,

pub fn get_copied(&self, key: &K) -> Option<V>

For cases when V is fixed-size, obtain a direct copy of V instead of V::ULE

pub fn iter_copied_values<'b>(
&'b self
) -> impl Iterator<Item = (&'b <K as ZeroMapKV<'a>>::GetType, V)>

Similar to Self::iter() except it returns a direct copy of the values instead of references to V::ULE, in cases when V is fixed-size

impl<'a, K, V> ZeroMapBorrowed<'a, K, V> where
K: ZeroMapKV<'a, Container = ZeroVec<'a, K>> + ?Sized,
V: ZeroMapKV<'a, Container = ZeroVec<'a, V>> + ?Sized,
K: AsULE + Copy + Ord + 'static,
V: AsULE + Copy + Ord + 'static,

pub fn iter_copied<'b: 'a>(&'b self) -> impl Iterator<Item = (K, V)> + 'b

Similar to Self::iter() except it returns a direct copy of the keys values instead of references to K::ULE and V::ULE, in cases when K and V are fixed-size