Trait zerovec::maps::MutableZeroVecLike

pub trait MutableZeroVecLike<'a, T: ?Sized>: ZeroVecLike<'a, T> {
    type OwnedType;
    fn zvl_insert(&mut self, index: usize, value: &T);
    fn zvl_remove(&mut self, index: usize) -> Self::OwnedType;
    fn zvl_replace(&mut self, index: usize, value: &T) -> Self::OwnedType;
    fn zvl_push(&mut self, value: &T);
    fn zvl_with_capacity(cap: usize) -> Self;
    fn zvl_clear(&mut self);
    fn zvl_reserve(&mut self, addl: usize);
    fn owned_as_t(o: &Self::OwnedType) -> &T;
}

Trait abstracting over ZeroVec and VarZeroVec, for use in ZeroMap. You should not be implementing or calling this trait directly.

This trait augments ZeroVecLike with methods allowing for mutation of the underlying vector for owned vector types.

Methods are prefixed with zvl_* to avoid clashes with methods on the types themselves

Associated Types

type OwnedType

The type returned by Self::remove() and Self::replace()

Required methods

fn zvl_insert(&mut self, index: usize, value: &T)

Insert an element at index

fn zvl_remove(&mut self, index: usize) -> Self::OwnedType

Remove the element at index (panicking if nonexistant)

fn zvl_replace(&mut self, index: usize, value: &T) -> Self::OwnedType

Replace the element at index with another one, returning the old element

fn zvl_push(&mut self, value: &T)

Push an element to the end of this vector

fn zvl_with_capacity(cap: usize) -> Self

Create a new, empty vector, with given capacity

fn zvl_clear(&mut self)

Remove all elements from the vector

fn zvl_reserve(&mut self, addl: usize)

Reserve space for addl additional elements

fn owned_as_t(o: &Self::OwnedType) -> &T

Convert an owned value to a borrowed T