Module zerovec::ule

Traits over unaligned little-endian data (ULE, pronounced "yule").

Structs

CharULE
A u8 array of little-endian data corresponding to a Unicode code point.

PlainOldULE
A u8 array of little-endian data with infallible conversions to and from &u8.

Traits

AsULE
A trait for any type that has a 1:1 mapping with an unaligned little-endian (ULE) type.

AsVarULE
A trait for any type that has a 1:1 mapping with an variable-width unaligned little-endian (VarULE) type.

ULE
Fixed-width, byte-aligned data that can be cast to and from a little-endian byte slice.

VarULE
Variable-width, byte-aligned data that can be cast to and from a little-endian byte slice.