zerovec::ule

Module niche

Structs

NichedOption

Optional type which uses NichedOptionULE<U,N> as ULE type.

Traits

NicheBytes

The ULE types implementing this trait guarantee that NicheBytes::NICHE_BIT_PATTERN can never occur as a valid byte representation of the type.

Unions

NichedOptionULE

ULE type for NichedOption<U,N> where U implements NicheBytes. The invalid bit pattern is used as the niche.