Struct zerovec::ule::UnvalidatedChar

#[repr(transparent)]
pub struct UnvalidatedChar(_);

A u8 array of little-endian data that is expected to be a Unicode scalar value, but is not validated as such.

Use this type instead of char when you want to deal with data that is expected to be valid Unicode scalar values, but you want control over when or if you validate that assumption.

Examples

use zerovec::ule::{RawBytesULE, UnvalidatedChar, ULE};
use zerovec::{ZeroSlice, ZeroVec};

// data known to be little-endian three-byte chunks of valid Unicode scalar values
let data = [0x68, 0x00, 0x00, 0x69, 0x00, 0x00, 0x4B, 0xF4, 0x01];
// ground truth expectation
let real = ['h', 'i', '๐'];

let chars: &ZeroSlice<UnvalidatedChar> = ZeroSlice::parse_byte_slice(&data).expect("invalid data length");
let parsed: Vec<_> = chars.iter().map(|c| unsafe { c.to_char_unchecked() }).collect();
assert_eq!(&parsed, &real);

let real_chars: ZeroVec<_> = real.iter().copied().map(UnvalidatedChar::from_char).collect();
let serialized_data = chars.as_bytes();
assert_eq!(serialized_data, &data);

Implementations

impl UnvalidatedChar

pub const fn from_char(c: char) -> Self

Create a UnvalidatedChar from a char.

Examples

use zerovec::ule::UnvalidatedChar;

let a = UnvalidatedChar::from_char('a');
assert_eq!(a.try_to_char().unwrap(), 'a');

pub fn try_to_char(self) -> Result<char, CharTryFromError>

Attempt to convert a UnvalidatedChar to a char.

Examples

use zerovec::ule::{AsULE, UnvalidatedChar};

let a = UnvalidatedChar::from_char('a');
assert_eq!(a.try_to_char(), Ok('a'));

let b = UnvalidatedChar::from_unaligned([0xFF, 0xFF, 0xFF].into());
assert!(matches!(b.try_to_char(), Err(_)));

pub fn to_char_lossy(self) -> char

Convert a UnvalidatedChar to a char', returning [char::REPLACEMENT_CHARACTER] if the UnvalidatedChar` does not represent a valid Unicode scalar value.

Examples

use zerovec::ule::{AsULE, UnvalidatedChar};

let a = UnvalidatedChar::from_unaligned([0xFF, 0xFF, 0xFF].into());
assert_eq!(a.to_char_lossy(), char::REPLACEMENT_CHARACTER);

pub unsafe fn to_char_unchecked(self) -> char

Convert a UnvalidatedChar to a char without checking that it is a valid Unicode scalar value.

Safety

The UnvalidatedChar must be a valid Unicode scalar value in little-endian order.

Examples

use zerovec::ule::UnvalidatedChar;

let a = UnvalidatedChar::from_char('a');
assert_eq!(unsafe { a.to_char_unchecked() }, 'a');