Trait zerovec::ule::EqULE

pub unsafe trait EqULE: AsULE { }

An EqULE type is one whose byte sequence equals the byte sequence of its ULE type on little-endian platforms. This enables certain performance optimizations, such as ZeroVec::try_from_slice.

Implementation safety

This trait is safe to implement if the type’s ULE (as defined by impl AsULE for T) has an equal byte sequence as the type itself on little-endian platforms; i.e., one where *const T can be cast to a valid *const T::ULE.

Object Safety

This trait is not object safe.