Trait zerovec::ule::VarULE

pub unsafe trait VarULE: 'static {
    type Error;
    fn parse_byte_slice(bytes: &[u8]) -> Result<&Self, Self::Error>;

    unsafe fn from_byte_slice_unchecked(bytes: &[u8]) -> &Self;

    fn as_byte_slice(&self) -> &[u8] { ... }
}

Variable-width, byte-aligned data that can be cast to and from a little-endian byte slice.

This trait is mostly for unsized types like str and [T]. It can be implemented on sized types, however it is much more preferable to use ULE for that purpose.

Safety

There must be no padding bytes involved in this type: Self::as_byte_slice() MUST return a slice of initialized bytes provided that Self is initialized.

VarULE::from_byte_slice_unchecked() must be implemented to return the same result as VarULE::parse_byte_slice() provided both are passed the same validly parsing byte slices. Both should return a pointer to the same region of memory that was passed in, covering that region completely.

VarULE::as_byte_slice() should return a slice that is the in-memory representation of Self, i.e. it should be just a pointer cast, and mem::size_of_val(self) == mem::size_of_val(self.as_byte_slice())

Equality invariant

A non-safety invariant is that if Self implements PartialEq, it must be logically equivalent to byte equality on .as_byte_slice(). If PartialEq does not imply byte-for-byte equality, then .parse_byte_slice() should return an error code for any values that are not in canonical form.

Failure to follow this invariant will cause surprising behavior in PartialEq, which may result in unpredictable operations on ZeroVec, VarZeroVec, and ZeroMap.

Associated Types

type Error

The error type to used by VarULE::parse_byte_slice()

Required methods

fn parse_byte_slice(bytes: &[u8]) -> Result<&Self, Self::Error>

Parses a byte slice, &[u8], and return it as &Self with the same lifetime.

If Self is not well-defined for all possible bit values, the bytes should be validated, and Self::Error should be returned if they are not valid.

Safety

Implementations of this method may involve unsafe{} blocks to cast the pointer to the correct type. It is up to the implementation to reason about the safety.

Implementors must return a pointer to the same region of memory if parsing succeeds.

unsafe fn from_byte_slice_unchecked(bytes: &[u8]) -> &Self

Takes a byte slice, &[u8], and return it as &self with the same lifetime, assuming that this byte slice has previously been run through VarULE::parse_byte_slice() with success.

There is no need to perform any validation here, this should almost always be a straight pointer cast.

Safety

Callers

Callers of this method must take care to ensure that bytes was previously passed through VarULE::parse_byte_slice() with success (and was not changed since then).

Implementors

This method must be implemented to return the same result as VarULE::parse_byte_slice().

Implementors must return a pointer to the same region of memory.

Implementations of this method may involve unsafe{} blocks to cast the pointer to the correct type. It is up to the implementation to reason about the safety, assuming the invariant above.

Provided methods

fn as_byte_slice(&self) -> &[u8]

Given &Self, returns a &[u8] with the same lifetime.

Safety

The implementation of this function should involve re-casting the pointer. It is up to the implementation to reason about the safety.

The default implementation already does this, however it can be overridden with a fully-safe method if possible.