zerovec::varzerovec

Module components

Structs

Index8

This is a VarZeroVecFormat that stores u8s in the index array, and a u8 for a length.

Index16

This is a VarZeroVecFormat that stores u16s in the index array, and a u16 for a length.

Index32

This is a VarZeroVecFormat that stores u32s in the index array, and a u32 for a length. Will have a larger data size, but will support large arrays without problems.

VarZeroSliceIter

An iterator over VarZeroSlice

VarZeroVecComponents

A more parsed version of VarZeroSlice. This type is where most of the VarZeroVec internal representation code lies.

Traits

VarZeroVecFormat

This trait allows switching between different possible internal representations of VarZeroVec.

Functions

compute_serializable_len

compute_serializable_len_without_length

get_serializable_bytes_non_empty

Collects the bytes for a VarZeroSlice into a Vec.

write_serializable_bytes

Writes the bytes for a VarZeroSlice into an output buffer.

write_serializable_bytes_without_length

Writes the bytes for a VarZeroLengthlessSlice into an output buffer. Usable for a VarZeroSlice if you first write the length bytes.