zerovec::varzerovec::components

Function write_serializable_bytes

pub fn write_serializable_bytes<T, A, F>(elements: &[A], output: &mut [u8])
where
    T: VarULE + ?Sized,
    A: EncodeAsVarULE<T>,
    F: VarZeroVecFormat,

Writes the bytes for a VarZeroSlice into an output buffer.

Every byte in the buffer will be initialized after calling this function.

Panics

Panics if the buffer is not exactly the correct length.