Struct zerovec::varzerovec::VarZeroSlice

#[repr(transparent)]
pub struct VarZeroSlice<T: ?Sized> { /* fields omitted */ }

A zero-copy “slice”, that works for unsized types, i.e. the zero-copy version of [T] where T is not Sized.

This behaves similarly to VarZeroVec<T>, however VarZeroVec<T> is allowed to contain owned data and as such is ideal for deserialization since most human readable serialization formats cannot unconditionally deserialize zero-copy.

This type can be nested within VarZeroVec<T> to allow for multi-level nested Vecs, for example the following code constructs the conceptual zero-copy equivalent of Vec<Vec<Vec<str>>>

use zerovec::VarZeroVec;
use zerovec::ZeroVec;
use zerovec::varzerovec::VarZeroSlice;
use zerovec::ule::*;
let strings_1: Vec<&str> = vec!["foo", "bar", "baz"];
let strings_2: Vec<&str> = vec!["twelve", "seventeen", "forty two"];
let strings_3: Vec<&str> = vec!["我", "喜歡", "烏龍茶"];
let strings_4: Vec<&str> = vec!["w", "ω", "文", "𑄃"];
let strings_12 = vec![strings_1.clone(), strings_2.clone()];
let strings_34 = vec![strings_3.clone(), strings_4.clone()];
let all_strings = vec![strings_12, strings_34];

let vzv_1: VarZeroVec<str> = VarZeroVec::from(&strings_1);
let vzv_2: VarZeroVec<str> = VarZeroVec::from(&strings_2);
let vzv_3: VarZeroVec<str> = VarZeroVec::from(&strings_3);
let vzv_4: VarZeroVec<str> = VarZeroVec::from(&strings_4);
let vzv_12 = VarZeroVec::from(&[vzv_1.as_slice(), vzv_2.as_slice()]);
let vzv_34 = VarZeroVec::from(&[vzv_3.as_slice(), vzv_4.as_slice()]);
let vzv_all = VarZeroVec::from(&[vzv_12.as_slice(), vzv_34.as_slice()]);

let reconstructed: Vec<Vec<Vec<String>>> = vzv_all.iter()
       .map(|v: &VarZeroSlice<VarZeroSlice<str>>| {
            v.iter().map(|x: &VarZeroSlice<_>| x.as_varzerovec().iter().map(|s| s.to_owned()).collect::<Vec<String>>())
             .collect::<Vec<_>>()
        }).collect::<Vec<_>>();
assert_eq!(reconstructed, all_strings);

let bytes = vzv_all.get_encoded_slice();
let vzv_from_bytes: VarZeroVec<VarZeroSlice<VarZeroSlice<str>>> = VarZeroVec::parse_byte_slice(bytes).unwrap();
assert_eq!(vzv_from_bytes, vzv_all);

Implementations

impl<T: VarULE + ?Sized> VarZeroSlice<T>

pub fn as_borrowed<'a>(&'a self) -> VarZeroVecBorrowed<'a, T>

Obtain a VarZeroVecBorrowed borrowing from the internal buffer

pub fn len(&self) -> usize

Get the number of elements in this vector

pub fn is_empty(&self) -> bool

Returns true if the vector contains no elements.

pub fn iter<'b>(&'b self) -> impl Iterator<Item = &'b T>

Obtain an iterator over VarZeroSlice’s elements

pub fn get(&self, idx: usize) -> Option<&T>

Get one of VarZeroSlice’s elements, returning None if the index is out of bounds

pub fn as_varzerovec<'a>(&'a self) -> VarZeroVec<'a, T>

Get this VarZeroSlice as a borrowed VarZeroVec

If you wish to repeatedly call methods on this VarZeroSlice, it is more efficient to perform this conversion first

impl<T> VarZeroSlice<T> where
T: VarULE,
T: ?Sized,
T: Ord,

pub fn binary_search(&self, x: &T) -> Result<usize, usize>

Binary searches a sorted VarZeroSlice<T> for the given element. For more information, see the primitive function binary_search.