Struct zerovec::varzerovec::VarZeroVec

pub struct VarZeroVec<'a, T>(_);

A zero-copy vector for variable-width types.

VarZeroVec<T> is designed as a drop-in replacement for Vec<T> in situations where it is desirable to borrow data from an unaligned byte slice, such as zero-copy deserialization, and where T’s data is variable-length (e.g. String)

T must implement AsVarULE, which is already implemented for [String] and Vec<T> where T implements ULE. It is also implemented on VarZeroVec<'static, T> if the inconvenience of converting to/from ULE is not desired. References are obtained as its AsVarULE::VarULE type, which in the case of [String] is str.

VarZeroVec<T> behaves much like [std::borrow::Cow], where it can be constructed from owned data but can also borrow from some buffer.

How it Works

VarZeroVec<T>, when used with non-human-readable serializers (like bincode), will serialize to a specially formatted list of bytes. The format is:

4 bytes for length (interpreted as a little-endian u32)

4 * length bytes of indices (interpreted as little-endian u32)

Remaining bytes for actual data

Each element in the indices array points to the starting index of its corresponding data part in the data list. The ending index can be calculated from the starting index of the next element (or the length of the slice if dealing with the last element).

Safety

VarZeroVec<T> is implemented with a fair amount of unsafe code, but is externally safe to use.

Example

use zerovec::VarZeroVec;

// The little-endian bytes correspond to the list of strings.
let strings = vec!["w".to_owned(), "ω".to_owned(), "文".to_owned(), "𑄃".to_owned()];
let bytes = &[
    4, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0,  3, 0, 0, 0,
    6, 0, 0, 0, 119, 207, 137, 230, 150, 135, 240, 145, 132, 131,
];

let zerovec: VarZeroVec<String> = VarZeroVec::parse_byte_slice(bytes)?;

assert_eq!(zerovec.get(2), Some("文"));
assert_eq!(zerovec, &*strings);

Here’s another example with Vec<T> and ZeroVec<'static, T>:

use zerovec::VarZeroVec;
use zerovec::ZeroVec;
use zerovec::ule::*;

// The little-endian bytes correspond to the list of integers.
let numbers: Vec<Vec<PlainOldULE<4>>> = vec![
    vec![12.into(), 25.into(), 38.into()],
    vec![39179.into(), 100.into()],
    vec![42.into(), 55555.into()],
    vec![12345.into(), 54321.into(), 9.into()],
];
let bytes = &[4, 0, 0, 0, 0, 0, 0, 0, 12, 0, 0, 0, 20, 0, 0, 0, 28, 0, 0, 0,
             12, 0, 0, 0, 25, 0, 0, 0, 38, 0, 0, 0, 11, 153, 0, 0, 100, 0,
             0, 0, 42, 0, 0, 0, 3, 217, 0, 0, 57, 48, 0, 0, 49, 212, 0, 0,
             9, 0, 0, 0];

let zerovec: VarZeroVec<Vec<PlainOldULE<4>>> = VarZeroVec::parse_byte_slice(bytes)?;
let zerovec2: VarZeroVec<ZeroVec<'static, u32>> = VarZeroVec::parse_byte_slice(bytes)?;

assert_eq!(zerovec.get(2).and_then(|v| v.get(1)), Some(&55555.into()));
assert_eq!(zerovec2.get(2).and_then(|v| v.get(1)), Some(&55555.into()));
assert_eq!(zerovec, &*numbers);
for (zv, v) in zerovec.iter().zip(numbers.iter()) {
    assert_eq!(zv, v);   
}

VarZeroVecs can be nested infinitely, see the docs of VarZeroVecULE for more information.

Implementations

impl<'a, T: AsVarULE> VarZeroVec<'a, T>

pub fn len(&self) -> usize

Get the number of elements in this vector

Example

let strings = vec!["foo".to_owned(), "bar".to_owned(),
                   "baz".to_owned(), "quux".to_owned()];
let bytes = VarZeroVec::get_serializable_bytes(&strings).unwrap();

let mut vec: VarZeroVec<String> = VarZeroVec::parse_byte_slice(&bytes)?;
assert_eq!(vec.len(), 4);

pub fn is_empty(&self) -> bool

Returns true if the vector contains no elements.

Examples

let strings: Vec<String> = vec![];
let bytes = VarZeroVec::get_serializable_bytes(&strings).unwrap();

let mut vec: VarZeroVec<String> = VarZeroVec::parse_byte_slice(&bytes)?;
assert!(vec.is_empty());

pub fn parse_byte_slice(slice: &'a [u8]) -> Result<Self, ParseErrorFor<T>>

Parse a VarZeroVec from a slice of the appropriate format

Slices of the right format can be obtained via VarZeroVec::get_serializable_bytes() or VarZeroVec::get_encoded_slice()

Example

let strings = vec!["foo".to_owned(), "bar".to_owned(),
                   "baz".to_owned(), "quux".to_owned()];
let bytes = VarZeroVec::get_serializable_bytes(&strings).unwrap();

let mut vec: VarZeroVec<String> = VarZeroVec::parse_byte_slice(&bytes)?;
assert_eq!(&vec[0], "foo");
assert_eq!(&vec[1], "bar");
assert_eq!(&vec[2], "baz");
assert_eq!(&vec[3], "quux");

pub fn iter<'b: 'a>(&'b self) -> impl Iterator<Item = &'b T::VarULE>

Obtain an iterator over VarZeroVec’s elements

Example

let strings = vec!["foo".to_owned(), "bar".to_owned(),
                   "baz".to_owned(), "quux".to_owned()];
let bytes = VarZeroVec::get_serializable_bytes(&strings).unwrap();
let mut vec: VarZeroVec<String> = VarZeroVec::parse_byte_slice(&bytes)?;

let mut iter_results: Vec<&str> = vec.iter().collect();
assert_eq!(iter_results[0], "foo");
assert_eq!(iter_results[1], "bar");
assert_eq!(iter_results[2], "baz");
assert_eq!(iter_results[3], "quux");

pub fn get(&self, idx: usize) -> Option<&T::VarULE>

Get one of VarZeroVec’s elements, returning None if the index is out of bounds

Example

let strings = vec!["foo".to_owned(), "bar".to_owned(),
                   "baz".to_owned(), "quux".to_owned()];
let bytes = VarZeroVec::get_serializable_bytes(&strings).unwrap();
let mut vec: VarZeroVec<String> = VarZeroVec::parse_byte_slice(&bytes)?;

let mut iter_results: Vec<&str> = vec.iter().collect();
assert_eq!(vec.get(0), Some("foo"));
assert_eq!(vec.get(1), Some("bar"));
assert_eq!(vec.get(2), Some("baz"));
assert_eq!(vec.get(3), Some("quux"));
assert_eq!(vec.get(4), None);

pub fn into_owned(self) -> VarZeroVec<'static, T> where
T: Clone,

Converts a borrowed ZeroVec to an owned ZeroVec. No-op if already owned.

Example

let strings = vec!["foo".to_owned(), "bar".to_owned(),
                   "baz".to_owned(), "quux".to_owned()];
let bytes = VarZeroVec::get_serializable_bytes(&strings).unwrap();
let mut vec: VarZeroVec<String> = VarZeroVec::parse_byte_slice(&bytes)?;

assert_eq!(vec.len(), 4);
// has 'static lifetime
let owned = vec.into_owned();

pub fn to_vec(&self) -> Vec<T> where
T: Clone,

Obtain an owned Vec<T> out of this

pub fn get_encoded_slice(&self) -> &[u8]

Obtain the internal encoded slice

This can be passed back to Self::parse_byte_slice()

pub fn get_serializable_bytes(elements: &[T]) -> Option<Vec<u8>>

For a slice of T, get a list of bytes that can be passed to parse_byte_slice to recoup the same data.

Returns None if the slice is too large to be represented in a list of bytes whose length fits in a u32.

Example

let strings = vec!["foo".to_owned(), "bar".to_owned(), "baz".to_owned()];
let bytes = VarZeroVec::get_serializable_bytes(&strings).unwrap();

let mut borrowed: VarZeroVec<String> = VarZeroVec::parse_byte_slice(&bytes)?;
assert_eq!(borrowed, &*strings);

pub fn is_owned(&self) -> bool

Return whether the VarZeroVec is operating on owned or borrowed data. VarZeroVec::into_owned() and VarZeroVec::make_mut() can be used to force it into an owned type

impl<'a, T> VarZeroVec<'a, T> where
T: AsVarULE,
T::VarULE: Ord,

pub fn binary_search(&self, x: &T::VarULE) -> Result<usize, usize>

Binary searches a sorted VarZeroVec<T> for the given element. For more information, see the primitive function binary_search.

Example

let strings = vec!["a".to_owned(), "b".to_owned(),
                   "f".to_owned(), "g".to_owned()];
let bytes = VarZeroVec::get_serializable_bytes(&strings).unwrap();
let mut vec: VarZeroVec<String> = VarZeroVec::parse_byte_slice(&bytes)?;

assert_eq!(vec.binary_search("f"), Ok(2));
assert_eq!(vec.binary_search("e"), Err(2));