Struct zerovec::varzerovec::VarZeroVecBorrowed

pub struct VarZeroVecBorrowed<'a, T: ?Sized> { /* fields omitted */ }

A fully-borrowed VarZeroVec. This type has the same internal buffer representation of VarZeroVec, making it cheaply convertible to VarZeroVec and VarZeroVecOwned; however, unlike those two, it cannot be mutated.

This is basically an &'a [u8] to a zero copy buffer, but split out into the buffer components. Logically this is capable of behaving as a &'a [T::VarULE], but since T::VarULE is unsized that type does not actually exist.

See VarZeroVecBorrowed::parse_byte_slice() for information on the internal invariants involved

Implementations

impl<'a, T: VarULE + ?Sized> VarZeroVecBorrowed<'a, T>

pub fn parse_byte_slice(slice: &'a [u8]) -> Result<Self, ParseErrorFor<T>>

Construct a new VarZeroVecBorrowed, checking invariants about the overall buffer size:

There must be either zero or at least four bytes (if four, this is the “length” parsed as a usize)

There must be at least 4*length + 4 bytes total, to form the the array indices of indices

indices[i]..indices[i+1] must index into a valid section of things, such that it parses to a T::VarULE

indices[len - 1]..things.len() must index into a valid section of things, such that it parses to a T::VarULE

pub unsafe fn from_bytes_unchecked(slice: &'a [u8]) -> Self

Construct a VarZeroVecBorrowed from a byte slice that has previously successfully returned a VarZeroVecBorrowed when passed to VarZeroVecBorrowed::parse_byte_slice(). Will return the same object as one would get from calling VarZeroVecBorrowed::parse_byte_slice().

Safety

The bytes must have previously successfully run through VarZeroVecBorrowed::parse_byte_slice()

pub fn len(self) -> usize

Get the number of elements in this vector

pub fn is_empty(self) -> bool

Returns true if the vector contains no elements.

pub fn entire_slice(self) -> &'a [u8]

Get a reference to the entire backing buffer of this vector

pub fn get(self, idx: usize) -> Option<&'a T>

Get the idx’th element out of this slice. Returns None if out of bounds.

pub fn iter(self) -> impl Iterator<Item = &'a T>

Create an iterator over the Ts contained in VarZeroVecBorrowed

pub fn to_vec(self) -> Vec<Box<T>>

impl<'a, T> VarZeroVecBorrowed<'a, T> where
T: VarULE,
T: ?Sized,
T: Ord,

pub fn binary_search(&self, needle: &T) -> Result<usize, usize>

Binary searches a sorted VarZeroVecBorrowed<T> for the given element. For more information, see the primitive function binary_search.