Struct zerovec::varzerovec::VarZeroVecOwned

pub struct VarZeroVecOwned<T: ?Sized> { /* fields omitted */ }

A fully-owned VarZeroVec. This type has no lifetime but has the same internal buffer representation of VarZeroVec, making it cheaply convertible to VarZeroVec and VarZeroVecBorrowed.

Implementations

impl<T: VarULE + ?Sized> VarZeroVecOwned<T>

pub fn new() -> Self

Construct an empty VarZeroVecOwned

pub fn from_borrowed(borrowed: VarZeroVecBorrowed<'_, T>) -> Self

Construct a VarZeroVecOwned from a VarZeroVecBorrowed by cloning the internal data

pub fn from_elements<A: EncodeAsVarULE<T>>(elements: &[A]) -> Self

Construct a VarZeroVecOwned from a list of elements

pub fn as_borrowed<'a>(&'a self) -> VarZeroVecBorrowed<'a, T>

Obtain a VarZeroVecBorrowed borrowing from the internal buffer

pub fn len(&self) -> usize

Get the number of elements in this vector

pub fn is_empty(&self) -> bool

Returns true if the vector contains no elements.

pub fn iter<'b>(&'b self) -> impl Iterator<Item = &'b T>

Obtain an iterator over VarZeroVecOwned’s elements

pub fn get(&self, idx: usize) -> Option<&T>

Get one of VarZeroVecOwned’s elements, returning None if the index is out of bounds

pub fn as_varzerovec<'a>(&'a self) -> VarZeroVec<'a, T>

Get this VarZeroVecOwned as a borrowed VarZeroVec

If you wish to repeatedly call methods on this VarZeroVecOwned, it is more efficient to perform this conversion first

pub fn clear(&mut self)

Empty the vector

pub fn to_vec(&self) -> Vec<Box<T>>

Convert this vector to a regular vector of boxed DSTs

pub fn entire_slice(&self) -> &[u8]

Get a reference to the entire backing buffer of this vector

pub fn push<A: EncodeAsVarULE<T> + ?Sized>(&mut self, element: &A)

Insert an element at the end of this vector

pub fn insert<A: EncodeAsVarULE<T> + ?Sized>(
&mut self,
index: usize,
element: &A
)

Insert an element at index idx

pub fn remove(&mut self, index: usize)

Remove the element at index idx

pub fn replace<A: EncodeAsVarULE<T> + ?Sized>(
&mut self,
index: usize,
element: &A
)

Replace the element at index idx with another

impl<T> VarZeroVecOwned<T> where
T: VarULE,
T: ?Sized,
T: Ord,

pub fn binary_search(&self, x: &T) -> Result<usize, usize>

Binary searches a sorted VarZeroVecOwned<T> for the given element. FoGeneralr more information, see the primitive function binary_search.