zerovec
Module vecs

This module contains additional utility types for working with ZeroVec and VarZeroVec. See their docs for more details on the general purpose of these types.

ZeroSlice and VarZeroSlice provide slice-like versions of the vector types for use behind references and in custom ULE types.

VarZeroVecOwned is a special owned/mutable version of VarZeroVec, allowing direct manipulation of the backing buffer.

Re-exports

pub use crate::zerovec::ZeroSlice;

pub use crate::zerovec::ZeroVec;

pub use crate::varzerovec::VarZeroSlice;

pub use crate::varzerovec::VarZeroVec;

Structs

Index8

This is a VarZeroVecFormat that stores u8s in the index array, and a u8 for a length.

Index16

This is a VarZeroVecFormat that stores u16s in the index array, and a u16 for a length.

Index32

This is a VarZeroVecFormat that stores u32s in the index array, and a u32 for a length. Will have a larger data size, but will support large arrays without problems.

VarZeroSliceIter

An iterator over VarZeroSlice

VarZeroVecOwned

A fully-owned VarZeroVec. This type has no lifetime but has the same internal buffer representation of VarZeroVec, making it cheaply convertible to VarZeroVec and VarZeroSlice.

ZeroSliceIter

An iterator over elements in a VarZeroVec

Traits

VarZeroVecFormat

This trait allows switching between different possible internal representations of VarZeroVec.

Type Aliases

VarZeroSlice16

VarZeroSlice32

VarZeroVec16

VarZeroVec32