Trait zerovec::map::ZeroVecLike

pub trait ZeroVecLike<'a, T> {
    type NeedleType: ?Sized;
    type GetType: ?Sized;
    fn binary_search(&self, k: &Self::NeedleType) -> Result<usize, usize>;

    fn get(&self, index: usize) -> Option<&Self::GetType>;

    fn insert(&mut self, index: usize, value: T);

    fn remove(&mut self, index: usize) -> T;

    fn replace(&mut self, index: usize, value: T) -> T;

    fn push(&mut self, value: T);

    fn len(&self) -> usize;

    fn new() -> Self;

    fn with_capacity(cap: usize) -> Self;

    fn clear(&mut self);

    fn reserve(&mut self, addl: usize);

    fn is_ascending(&self) -> bool;

    fn is_empty(&self) -> bool { ... }
}

Trait abstracting over ZeroVec and VarZeroVec, for use in ZeroMap. You should not be implementing or calling this trait directly.

Associated Types

type NeedleType: ?Sized

The type received by Self::binary_search()

type GetType: ?Sized

The type returned by Self::get()

Loading content...
Required methods

fn binary_search(&self, k: &Self::NeedleType) -> Result<usize, usize>

Search for a key in a sorted vector, returns Ok(index) if found, returns Err(insert_index) if not found, where insert_index is the index where it should be inserted to maintain sort order.

fn get(&self, index: usize) -> Option<&Self::GetType>

Get element at index

fn insert(&mut self, index: usize, value: T)

Insert an element at index

fn remove(&mut self, index: usize) -> T

Remove the element at index (panicking if nonexistant)

fn replace(&mut self, index: usize, value: T) -> T

Replace the element at index with another one, returning the old element

fn push(&mut self, value: T)

Push an element to the end of this vector

fn len(&self) -> usize

The length of this vector

fn new() -> Self

Create a new, empty vector

fn with_capacity(cap: usize) -> Self

Create a new, empty vector, with given capacity

fn clear(&mut self)

Remove all elements from the vector

fn reserve(&mut self, addl: usize)

Reserve space for addl additional elements

fn is_ascending(&self) -> bool

Check if this vector is in ascending order according to Ts Ord impl

Loading content...
Provided methods

fn is_empty(&self) -> bool

Check if this vector is empty

Loading content...