zerovec

Module maps

This module contains additional utility types and traits for working with ZeroMap and ZeroMap2d. See their docs for more details on the general purpose of these types.

ZeroMapBorrowed and ZeroMap2dBorrowed are versions of ZeroMap and ZeroMap2d that can be used when you wish to guarantee that the map data is always borrowed, leading to relaxed lifetime constraints.

The ZeroMapKV trait is required to be implemented on any type that needs to be used within a map type. ZeroVecLike and MutableZeroVecLike are traits used in the internal workings of the map types, and should typically not be used or implemented by users of this crate.

Re-exports

pub use crate::map::ZeroMap;

pub use crate::map2d::ZeroMap2d;

Structs

ZeroMap2dBorrowed

A borrowed-only version of ZeroMap2d

ZeroMap2dCursor

An intermediate state of queries over ZeroMap2d and ZeroMap2dBorrowed.

ZeroMapBorrowed

A borrowed-only version of ZeroMap

Traits

MutableZeroVecLike

Trait abstracting over ZeroVec and VarZeroVec, for use in ZeroMap. You should not be implementing or calling this trait directly.

ZeroMapKV

Trait marking types which are allowed to be keys or values in ZeroMap.

ZeroVecLike

Trait abstracting over ZeroVec and VarZeroVec, for use in ZeroMap. You should not be implementing or calling this trait directly.