zerovec

Struct VarZeroCow

pub struct VarZeroCow<'a, V: ?Sized> { /* private fields */ }

Copy-on-write type that efficiently represents VarULE types as their bitstream representation.

The primary use case for VarULE types is the ability to store complex variable-length datastructures inside variable-length collections like crate::VarZeroVec.

Underlying this ability is the fact that VarULE types can be efficiently represented as a flat bytestream.

In zero-copy cases, sometimes one wishes to unconditionally use this bytestream representation, for example to save stack size. A struct with five Cow<'a, str>s is not as stack-efficient as a single Cow containing the bytestream representation of, say, Tuple5VarULE<str, str, str, str, str>.

This type helps in this case: It is logically a Cow<'a, V>, with some optimizations, that is guaranteed to serialize as a byte stream in machine-readable scenarios.

During human-readable serialization, it will fall back to the serde impls on V, which ought to have a human-readable variant.

Implementations

impl<'a, V: VarULE + ?Sized> VarZeroCow<'a, V>

pub fn parse_bytes(bytes: &'a [u8]) -> Result<Self, UleError>

Construct from a slice. Errors if the slice doesn’t represent a valid V

pub const unsafe fn from_bytes_unchecked(bytes: &'a [u8]) -> Self

Construct from a slice that is known to represent a valid V

Safety

bytes must be a valid V, i.e. it must successfully pass through V::parse_bytes() or V::validate_bytes().

pub fn new_borrowed(val: &'a V) -> Self

Construct a new borrowed version of this

impl<'a, V: ?Sized> VarZeroCow<'a, V>

pub fn is_owned(&self) -> bool

Whether or not this is owned

pub fn as_bytes(&self) -> &[u8]

Get the byte representation of this type

Is also always a valid V and can be passed to V::from_bytes_unchecked()