Module zerovec::ule

Traits over unaligned little-endian data (ULE, pronounced “yule”).

Modules

custom

This module contains utilities that help define custom VarULE types, especially those using complex custom dynamically sized types.

Structs

CharULE

A u8 array of little-endian data corresponding to a Unicode code point.

PairULE

We do not have guarantees for the layouts of tuples, so we must define a custom ULE type for pairs. This could potentially be generalized for larger tuples if necessary

RawBytesULE

A u8 array of little-endian data with infallible conversions to and from &u8.

Enums

ZeroVecError

A generic error type to be used for decoding slices of ULE types

Traits

AsULE

A trait for any type that has a 1:1 mapping with an unaligned little-endian (ULE) type.

EqULE

An EqULE type is one whose byte sequence equals the byte sequence of its ULE type on little-endian platforms. This enables certain performance optimizations, such as ZeroVec::try_from_slice.

SliceAsULE

A trait for a type where aligned slices can be cast to unaligned slices.

ULE

Fixed-width, byte-aligned data that can be cast to and from a little-endian byte slice.

VarULE

Variable-width, byte-aligned data that can be cast to and from a little-endian byte slice.