Derive Macro zerovec::ule::VarULE

#[derive(VarULE)]

Custom derive for VarULE

This can be attached to structs containing only ULE types with one VarULE type at the end.

Most of the time, it is recommended one use #[make_varule] instead of defining a custom VarULE type. Full docs for this proc macro can be found on the zerovec crate.