zerovec::ule

Module multi

Structs

BlankSliceEncoder 🔒

This lets us conveniently use the EncodeAsVarULE functionality to create VarZeroVec<[u8]>s that have the right amount of space for elements without having to duplicate any unsafe code

MultiFieldsULE

This type is used by the custom derive to represent multiple VarULE fields packed into a single end-of-struct field. It is not recommended to use this type directly.