zerovec::ule

Struct MultiFieldsULE

pub struct MultiFieldsULE<const LEN: usize, Format: VarZeroVecFormat>(/* private fields */);

This type is used by the custom derive to represent multiple VarULE fields packed into a single end-of-struct field. It is not recommended to use this type directly, use Tuple2VarULE etc instead.

Logically, consider it to be (, , , ..) where `` etc are potentially different VarULE types.

Internally, it is represented by a VarZeroSlice without the length part.

Implementations

impl<const LEN: usize, Format: VarZeroVecFormat> MultiFieldsULE<LEN, Format>

pub fn compute_encoded_len_for(lengths: [usize; LEN]) -> usize

Compute the amount of bytes needed to support elements with lengths lengths

pub fn new_from_lengths_partially_initialized<'a>( lengths: [usize; LEN], output: &'a mut [u8], ) -> &'a mut Self

Construct a partially initialized MultiFieldsULE backed by a mutable byte buffer

pub unsafe fn set_field_at<T: VarULE + ?Sized, A: EncodeAsVarULE<T> + ?Sized>( &mut self, idx: usize, value: &A, )

Given a buffer of size obtained by Self::compute_encoded_len_for(), write element A to index idx

Safety

idx must be in range

T must be the appropriate type expected by the custom derive in this usage of this type

pub unsafe fn validate_field<T: VarULE + ?Sized>( &self, index: usize, ) -> Result<(), UleError>

Validate field at index to see if it is a valid T VarULE type

Safety

index must be in range

pub unsafe fn get_field<T: VarULE + ?Sized>(&self, index: usize) -> &T

Get field at index as a value of type T

Safety

index must be in range

Element at index must have been created with the VarULE type T

pub unsafe fn from_bytes_unchecked(bytes: &[u8]) -> &Self

Construct from a byte slice

Safety

byte slice must be a valid VarZeroLengthlessSlice<u8, Format> with length LEN

pub fn as_bytes(&self) -> &[u8]

Get the bytes behind this value