zerovec::varzerovec

Module components

Structs

Index16

This is a VarZeroVecFormat that stores u16s in the index array. Will have a smaller data size, but it’s more likely for larger arrays to be unrepresentable (and error on construction)

Index32

This is a VarZeroVecFormat that stores u32s in the index array. Will have a larger data size, but will support large arrays without problems.

VarZeroVecComponents

A more parsed version of VarZeroSlice. This type is where most of the VarZeroVec internal representation code lies.

Constants

LENGTH_WIDTH 🔒

MAX_INDEX 🔒

MAX_LENGTH 🔒

METADATA_WIDTH 🔒

Traits

VarZeroVecFormat

This trait allows switching between different possible internal representations of VarZeroVec.

Functions

compute_serializable_len

get_serializable_bytes_non_empty

Collects the bytes for a VarZeroSlice into a Vec.

write_serializable_bytes

Writes the bytes for a VarZeroSlice into an output buffer.