Module zerovec::vecs

This module contains additional utility types for working with ZeroVec and VarZeroVec. See their docs for more details on the general purpose of these types.

ZeroSlice and VarZeroSlice provide slice-like versions of the vector types for use behind references and in custom ULE types.

VarZeroVecOwned is a special owned/mutable version of VarZeroVec, allowing direct manipulation of the backing buffer.

Re-exports

pub use crate::zerovec::ZeroSlice;

pub use crate::zerovec::ZeroVec;

pub use crate::varzerovec::VarZeroSlice;

pub use crate::varzerovec::VarZeroVec;

Structs

VarZeroVecOwned

A fully-owned VarZeroVec. This type has no lifetime but has the same internal buffer representation of VarZeroVec, making it cheaply convertible to VarZeroVec and VarZeroSlice.