zerovec::vecs

Struct FlexZeroVecOwned

pub struct FlexZeroVecOwned(/* private fields */);

The fully-owned variant of FlexZeroVec. Contains all mutation methods.

Implementations

impl FlexZeroVecOwned

pub fn new_empty() -> Self

Creates a new FlexZeroVecOwned with zero elements.

pub fn from_slice(other: &FlexZeroSlice) -> FlexZeroVecOwned

Creates a FlexZeroVecOwned from a FlexZeroSlice.

pub fn as_slice(&self) -> &FlexZeroSlice

Obtains this FlexZeroVecOwned as a FlexZeroSlice.

pub fn into_flexzerovec(self) -> FlexZeroVec<'static>

Converts this FlexZeroVecOwned into a FlexZeroVec::Owned.

pub fn clear(&mut self)

Clears all values out of this FlexZeroVecOwned.

pub fn push(&mut self, item: usize)

Appends an item to the end of the vector.

Panics

Panics if inserting the element would require allocating more than usize::MAX bytes.

Examples

use zerovec::vecs::FlexZeroVec;

let mut zv: FlexZeroVec = [22, 44, 66].iter().copied().collect();
zv.to_mut().push(33);
assert_eq!(zv.to_vec(), vec![22, 44, 66, 33]);

pub fn insert(&mut self, index: usize, item: usize)

Inserts an element into the middle of the vector.

Caution: Both arguments to this function are of type usize. Please be careful to pass the index first followed by the value second.

Panics

Panics if index > len.

Panics if inserting the element would require allocating more than usize::MAX bytes.

Examples

use zerovec::vecs::FlexZeroVec;

let mut zv: FlexZeroVec = [22, 44, 66].iter().copied().collect();
zv.to_mut().insert(2, 33);
assert_eq!(zv.to_vec(), vec![22, 44, 33, 66]);

pub fn insert_sorted(&mut self, item: usize)

Inserts an element into an ascending sorted vector at a position that keeps the vector sorted.

Panics

Panics if inserting the element would require allocating more than usize::MAX bytes.

Examples

use zerovec::vecs::FlexZeroVecOwned;

let mut fzv = FlexZeroVecOwned::new_empty();
fzv.insert_sorted(10);
fzv.insert_sorted(5);
fzv.insert_sorted(8);

assert!(Iterator::eq(fzv.iter(), [5, 8, 10].iter().copied()));

pub fn remove(&mut self, index: usize) -> usize

Removes and returns the element at the specified index.

Panics

Panics if index >= len.

Examples

use zerovec::vecs::FlexZeroVec;

let mut zv: FlexZeroVec = [22, 44, 66].iter().copied().collect();
let removed_item = zv.to_mut().remove(1);
assert_eq!(44, removed_item);
assert_eq!(zv.to_vec(), vec![22, 66]);

pub fn pop_sorted(&mut self) -> usize

Removes and returns the last element from an ascending sorted vector.

If the vector is not sorted, use FlexZeroVecOwned::remove() instead. Calling this function would leave the FlexZeroVec in a safe, well-defined state; however, information may be lost and/or the equality invariant might not hold.

Panics

Panics if self.is_empty().

Examples

use zerovec::vecs::FlexZeroVec;

let mut zv: FlexZeroVec = [22, 44, 66].iter().copied().collect();
let popped_item = zv.to_mut().pop_sorted();
assert_eq!(66, popped_item);
assert_eq!(zv.to_vec(), vec![22, 44]);

Calling this function on a non-ascending vector could cause surprising results:

use zerovec::vecs::FlexZeroVec;

let mut zv1: FlexZeroVec = [444, 222, 111].iter().copied().collect();
let popped_item = zv1.to_mut().pop_sorted();
assert_eq!(111, popped_item);

// Oops!
assert_eq!(zv1.to_vec(), vec![188, 222]);