zerovec::vecs

Struct VarZeroVecOwned

pub struct VarZeroVecOwned<T: ?Sized, F = Index16> {
    marker: PhantomData<(Box<T>, F)>,
    entire_slice: Vec<u8>,
}

A fully-owned VarZeroVec. This type has no lifetime but has the same internal buffer representation of VarZeroVec, making it cheaply convertible to VarZeroVec and VarZeroSlice.

The F type parameter is a VarZeroVecFormat (see its docs for more details), which can be used to select the precise format of the backing buffer with various size and performance tradeoffs. It defaults to Index16.

Fields
marker: PhantomData<(Box<T>, F)>entire_slice: Vec<u8>
Implementations

impl<T: VarULE + ?Sized, F> VarZeroVecOwned<T, F>

pub fn new() -> Self

Construct an empty VarZeroVecOwned

impl<T: VarULE + ?Sized, F: VarZeroVecFormat> VarZeroVecOwned<T, F>

pub fn from_slice(slice: &VarZeroSlice<T, F>) -> Self

Construct a VarZeroVecOwned from a VarZeroSlice by cloning the internal data

pub fn try_from_elements<A>(elements: &[A]) -> Result<Self, &'static str>
where A: EncodeAsVarULE<T>,

Construct a VarZeroVecOwned from a list of elements

pub fn as_slice(&self) -> &VarZeroSlice<T, F>

Obtain this VarZeroVec as a VarZeroSlice

pub(crate) fn with_capacity(capacity: usize) -> Self

Try to allocate a buffer with enough capacity for capacity elements. Since T can take up an arbitrary size this will just allocate enough space for 4-byte Ts

pub(crate) fn reserve(&mut self, capacity: usize)

Try to reserve space for capacity elements. Since T can take up an arbitrary size this will just allocate enough space for 4-byte Ts

unsafe fn element_position_unchecked(&self, idx: usize) -> usize

Get the position of a specific element in the data segment.

If idx == self.len(), it will return the size of the data segment (where a new element would go).

Safety

idx <= self.len() and self.as_encoded_bytes() is well-formed.

unsafe fn element_range_unchecked(&self, idx: usize) -> Range<usize>

Get the range of a specific element in the data segment.

Safety

idx < self.len() and self.as_encoded_bytes() is well-formed.

unsafe fn set_len(&mut self, len: usize)

Set the number of elements in the list without any checks.

Safety

No safe functions may be called until self.as_encoded_bytes() is well-formed.

fn index_range(index: usize) -> Range<usize>

unsafe fn index_data(&self, index: usize) -> &F::RawBytes

Return the raw bytes representing the given index.

Safety

The index must be valid, and self.as_encoded_bytes() must be well-formed

unsafe fn index_data_mut(&mut self, index: usize) -> &mut F::RawBytes

Return the mutable slice representing the given index.

Safety

The index must be valid. self.as_encoded_bytes() must have allocated space for this index, but need not have its length appropriately set.

unsafe fn shift_indices(&mut self, starting_index: usize, amount: i32)

Shift the indices starting with and after starting_index by the provided amount.

Safety

Adding amount to each index after starting_index must not result in the slice from becoming malformed. The length of the slice must be correctly set.

pub fn as_varzerovec<'a>(&'a self) -> VarZeroVec<'a, T, F>

Get this VarZeroVecOwned as a borrowed VarZeroVec

If you wish to repeatedly call methods on this VarZeroVecOwned, it is more efficient to perform this conversion first

pub fn clear(&mut self)

Empty the vector

pub fn into_bytes(self) -> Vec<u8>

Consume this vector and return the backing buffer

unsafe fn shift( &mut self, index: usize, new_size: usize, shift_type: ShiftType, ) -> &mut [u8]

Invalidate and resize the data at an index, optionally inserting or removing the index. Also updates affected indices and the length. Returns a slice to the new element data - it doesn’t contain uninitialized data but its value is indeterminate.

Safety

index must be a valid index, or, if shift_type == ShiftType::Insert, index == self.len() is allowed.

new_size musn’t result in the data segment growing larger than F::MAX_VALUE.

fn verify_integrity(&self) -> bool

Checks the internal invariants of the vec to ensure safe code will not cause UB. Returns whether integrity was verified.

Note: an index is valid if it doesn’t point to data past the end of the slice and is less than or equal to all future indices. The length of the index segment is not part of each index.

pub fn push<A: EncodeAsVarULE<T> + ?Sized>(&mut self, element: &A)

Insert an element at the end of this vector

pub fn insert<A: EncodeAsVarULE<T> + ?Sized>( &mut self, index: usize, element: &A, )

Insert an element at index idx

pub fn remove(&mut self, index: usize)

Remove the element at index idx

pub fn replace<A: EncodeAsVarULE<T> + ?Sized>( &mut self, index: usize, element: &A, )

Replace the element at index idx with another