Struct zerovec::vecs::VarZeroVecOwned

pub struct VarZeroVecOwned<T: ?Sized> { /* fields omitted */ }

A fully-owned VarZeroVec. This type has no lifetime but has the same internal buffer representation of VarZeroVec, making it cheaply convertible to VarZeroVec and VarZeroSlice.

Implementations

impl<T: VarULE + ?Sized> VarZeroVecOwned<T>

pub fn new() -> Self

Construct an empty VarZeroVecOwned

pub fn from_slice(slice: &VarZeroSlice<T>) -> Self

Construct a VarZeroVecOwned from a VarZeroSlice by cloning the internal data

pub fn try_from_elements<A>(elements: &[A]) -> Result<Self, &'static str> where
A: EncodeAsVarULE<T>,

Construct a VarZeroVecOwned from a list of elements

pub fn as_slice(&self) -> &VarZeroSlice<T>

Obtain this VarZeroVec as a VarZeroSlice

pub fn as_varzerovec<'a>(&'a self) -> VarZeroVec<'a, T>

Get this VarZeroVecOwned as a borrowed VarZeroVec

If you wish to repeatedly call methods on this VarZeroVecOwned, it is more efficient to perform this conversion first

pub fn clear(&mut self)

Empty the vector

pub fn into_bytes(self) -> Vec<u8>

Consume this vector and return the backing buffer

pub fn push<A: EncodeAsVarULE<T> + ?Sized>(&mut self, element: &A)

Insert an element at the end of this vector

pub fn insert<A: EncodeAsVarULE<T> + ?Sized>(
&mut self,
index: usize,
element: &A
)

Insert an element at index idx

pub fn remove(&mut self, index: usize)

Remove the element at index idx

pub fn replace<A: EncodeAsVarULE<T> + ?Sized>(
&mut self,
index: usize,
element: &A
)

Replace the element at index idx with another