Trait zerovec::vecs::VarZeroVecFormat

pub unsafe trait VarZeroVecFormat: 'static + Sized { }

This trait allows switching between different possible internal representations of VarZeroVec.

Currently this crate supports two formats: Index16 and Index32, with Index16 being the default for all VarZeroVec types unless explicitly specified otherwise.

Do not implement this trait, its internals may be changed in the future, and all of its associated items are hidden from the docs.